PCME-DAA/requirements.txt
adamp==0.3.0
apex==0.1
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cxxfilt==0.3.0
cycler==0.11.0
Cython==0.29.28
filelock==3.6.0
fire==0.4.0
fonttools==4.29.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
gspread==5.1.1
httplib2==0.20.4
huggingface-hub==0.5.1
idna==3.3
iniconfig==1.1.1
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
munch==2.5.0
nltk==3.7
numpy==1.22.2
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rouge==1.0.1
rsa==4.8
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.12.1
tomli==2.0.1
torch==1.9.0+cu111
torchaudio==0.9.0
torchtext==0.10.0
torchvision==0.10.0+cu111
tqdm==4.62.3
transformers==4.18.0
typing_extensions==4.1.1
ujson==5.1.0
urllib3==1.26.8
